Change int(num) to float(num) for more accurate math

+++ probSet0.py
@@ -3,8 +3,8 @@ import numpy
 x = input("Please enter a number for x: ")
 y = input("Please enter a number for y: ")
 
-x = int(x)
-y = int(y)
+x = float(x)
+y = float(y)
 
 print("x**y = ", x**y)
 print("log(x) = ", numpy.log2(x))
